// src-ui/core/ipc/generated/domain/HttpSseServerConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessagePostFormat } from "./MessagePostFormat";

export type HttpSseServerConfig = { host: string, port: number, enableCors: boolean, enableWebsocket: boolean, reportSelfMessage: boolean, enable: boolean, name: string, messagePostFormat: MessagePostFormat, token: string, debug: boolean, };