// src-ui/core/ipc/generated/domain/WebsocketClientConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessagePostFormat } from "./MessagePostFormat";
import type { WsRole } from "./WsRole";

export type WebsocketClientConfig = { url: string, reportSelfMessage: boolean, heartInterval: number, reconnectInterval: number, role: WsRole, enable: boolean, name: string, messagePostFormat: MessagePostFormat, token: string, debug: boolean, };